requirements.txt
fastapi~=0.110.2
pydantic~=2.7.0
sqlalchemy~=2.0.29
python-decouple~=3.8
gunicorn~=22.0.0
uvicorn~=0.29.0
starlette~=0.37.2

